# apis/openapi.yaml
openapi: 3.0.3
info:
  title: Order Management API
  version: "2.0.0"
  description: >
    The missing upstream Open Host Service for the warehouse-systems
    platform: it owns **Order** and **OrderLine** as first-class, validated
    aggregates. Before this bounded context existed, "an order" was just an
    unowned, unvalidated string — `OrderRef` / `DemandRef` / `Reference` —
    independently reinvented by three different services. Here the OrderId
    is a real identity, and it is the value sent downstream as
    inventory-storage's `demandRef`.


    The lifecycle is intake -> allocation -> release, with cancellation
    legal up to the release boundary — but as of this redesign, a caller
    only ever expresses ONE intent: **place an order**
    (`POST /orders`). Allocation and release are internal saga steps this
    service performs automatically, right after intake, and again on
    `retry-allocation` — they are no longer public REST verbs a caller
    drives by hand. **Allocation** reserves stock per line through
    inventory-storage's published `POST /reservations`; a `409` from that
    service is the business fact "no usable stock", which backorders that
    one line, while any other failure fails the whole allocation pass
    rather than silently recording a backorder. **Release** is no longer a
    synchronous call to any Supplier — it is announced as a fact on the
    enriched `OrderAllocated` / `OrderPartiallyAllocated` Kafka integration
...


...
    summarises.
...
servers:
...
paths:
...
  /orders:
    post:
      operationId: receiveOrder
      summary: Receive a new order — allocates and releases automatically
      description: >
        Intake — the ReceiveOrder use case. Validates every requested line
        (SKU non-empty, quantity > 0), mints an `OrderId`, and persists the
        order in `Received` status, publishing `OrderReceived`
        unconditionally.


        Immediately afterward, in the SAME call, this service attempts
        allocation-then-release: it reserves stock per line via
        inventory-storage's `POST /reservations` (see the Allocation tag),
        and for every line that ends up `Allocated` and eligible under BR3
        (see below), releases it — announced via the enriched
        `OrderAllocated` / `OrderPartiallyAllocated` Kafka events, never a
        second synchronous call. If that implicit allocation attempt hits
        a hard (non-business) failure, this endpoint still returns `201`:
        the order genuinely was received, and a caller expressing "place
        an order" must never be told that failed just because an internal,
        best-effort next step could not complete. The returned body
        reflects whatever the allocation pass actually achieved.


        `allowPartialShipment` defaults to `false` — ship-complete (BR3).
        Every line's `pathId` is always the internal default (`pick`) — a
        caller placing an order has no business supplying
        wes-work-planning's process-path vocabulary at intake; `pathId` is
        visible on every line in the response, but not settable on the
        request.
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiveOrderRequest"
            example:
              allowPartialShipment: false
              lines:
                - sku: SKU-1
                  quantity: 2
                - sku: SKU-2
                  quantity: 1
                  giftWrap: true
      responses:
        "201":
          description: >
            Order received. Reflects whatever the immediate
            allocation-then-release attempt achieved — `Received`,
            `Backordered`, `PartiallyAllocated`/`PartiallyReleased`,
            `Allocated`, or `Released` — a caller inspects `status`/line
            `status` for the outcome rather than assuming `Received`.
          headers:
            Location:
              schema:
                type: string
              description: Canonical URI of the created order.
              example: /orders/ord-a1b2c3d4-0000-0000-0000-000000000001
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
              example:
                id: ord-a1b2c3d4-0000-0000-0000-000000000001
                status: Released
                allowPartialShipment: false
                releaseOnAllocation: true
                promiseDate: "2026-08-27T09:00:00Z"
                lines:
                  - lineNo: 1
                    sku: SKU-1
                    quantity: 2
                    pathId: pick
                    giftWrap: false
                    status: Released
                    reservationId: res-a1b2c3d4-0000-0000-0000-000000000009
                  - lineNo: 2
                    sku: SKU-2
                    quantity: 1
                    pathId: pick
                    giftWrap: true
                    status: Released
                    reservationId: res-a1b2c3d4-0000-0000-0000-000000000010
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          description: A line quantity is not greater than zero.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://errors.order-management.warehouse-systems.dev/non-positive-quantity
                title: Quantity must be greater than zero
                status: 422
                detail: quantity must be greater than zero
                instance: /orders
        "500":
          $ref: "#/components/responses/InternalError"
  /orders/{id}:
    get:
      operationId: getOrder
      summary: Read an order's current state
      description: >
        The GetOrder use case. Returns the order with every line's current
        status, its reservation reference where allocated, and the promise
        date once allocation has computed one. The order-level `status` in
        this response is derived from the line statuses at read time — it
        is not a stored field.
      tags: [Orders]
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: The order's current state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
              example:
                id: ord-a1b2c3d4-0000-0000-0000-000000000001
                status: Allocated
                allowPartialShipment: false
                releaseOnAllocation: true
                promiseDate: "2026-08-27T09:00:00Z"
                lines:
                  - lineNo: 1
                    sku: SKU-1
                    quantity: 2
                    pathId: pick
                    giftWrap: false
                    status: Allocated
                    reservationId: res-a1b2c3d4-0000-0000-0000-000000000009
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/OrderNotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: cancelOrder
      summary: Cancel an order and revoke its reservations
      description: >
        The CancelOrder use case. Revokes every allocated line's
        reservation on inventory-storage via
        `DELETE /reservations/{id}`, then cancels every line. Publishes
        `OrderCancelled`.


        **BR6 — the cancellation boundary:** this is legal ONLY while no
        line has reached `Released`. The boundary is checked BEFORE any
        reservation is revoked, so a rejected cancellation leaves
        inventory-storage completely untouched. Once any line is released,
        v1 does NOT claw back the released work — a deliberate, documented
        gap (see ADR 0004), not an oversight.


        A failed revoke fails the whole cancellation and leaves the order
        untouched, so retrying is safe.
      tags: [Orders]
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "204":
          description: Order cancelled; every allocated reservation was revoked.
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/OrderNotFound"
        "409":
          description: At least one line has already been released (BR6).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://errors.order-management.warehouse-systems.dev/order-already-released
                title: Order already has released lines and can no longer be cancelled
                status: 409
                detail: order already has released lines and can no longer be cancelled
                instance: /orders/ord-a1b2c3d4-0000-0000-0000-000000000001
        "503":
          $ref: "#/components/responses/DownstreamUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"
  /orders/{id}/release:
    post:
      operationId: releaseHeldOrder
      summary: Release an order held at intake
      description: >
        The ReleaseHeldOrder use case (ADR 0020). Commits an order received
        with `releaseOnAllocation: false` to the floor: its already-allocated
        lines are released through the same BR3-gated flow every other
        release uses, the promise is recomputed by the same policy, and the
        same `OrderAllocated`/`OrderReleased` events fire with the same
        per-line payload.


        IDEMPOTENT for a HELD order: releasing one whose lines are already
        released returns `200` with no state change and no event. A caller
        working against an external fill-or-kill deadline cannot distinguish
        "my release was lost" from "my release failed", so retrying must be
        safe.


        An order that was never held has nothing to release on demand — its
        lines release themselves at allocation — so calling this on one is
        rejected with `409` rather than silently succeeding, since it means
        the caller has the wrong order id or the wrong expectation. This
        holds even when that order has no allocated lines of its own (a
        backordered order, say): "never held" and "already released" are
        different answers and must not collapse into one.
      tags: [Allocation]
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: The order after release (or unchanged, if already released).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          $ref: "#/components/responses/OrderNotFound"
        "409":
          description: >
            This order was not held at intake, so it has nothing to release
            on demand.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://errors.order-management.warehouse-systems.dev/order-not-held
                title: Order was not held at intake
                status: 409
                detail: order was not held at intake
                instance: /orders/ord-a1b2c3d4-0000-0000-0000-000000000001/release
        "503":
          $ref: "#/components/responses/DownstreamUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"
  /orders/{id}/retry-allocation:
    post:
      operationId: retryAllocation
      summary: Re-attempt allocation for backordered lines, then release
      description: >
        The RetryAllocation use case. Re-attempts allocation for the
        order's `Backordered` lines and ONLY those lines — it is the single
        sanctioned route from `Backordered` back to `Allocated`, and it is
        what unblocks a ship-complete order held back by BR3. UNLIKE the
        implicit allocation attempt folded into `POST /orders`, this is an
        explicit, on-purpose operator/recovery action: a hard (non-409)
        failure here fails the call with `503` rather than being swallowed
        — invoking this endpoint IS asking for allocation right now.


        A line still short of stock stays `Backordered` (no error). The
        same fail-closed rule as intake's implicit allocation applies to
        any non-409 failure. An order with no backordered line at all is
        rejected with `409`: retrying an order that is not backordered is
        a caller mistake, not a no-op.


        On success, this ALSO attempts release in the same call — the same
        allocation-then-release flow `POST /orders` uses — announced via
        the enriched `OrderAllocated` / `OrderPartiallyAllocated` Kafka
        events rather than any synchronous call to a Supplier.
      tags: [Allocation]
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: The order after the retry-then-release pass.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
              example:
                id: ord-a1b2c3d4-0000-0000-0000-000000000001
                status: Released
                allowPartialShipment: false
                releaseOnAllocation: true
                promiseDate: "2026-08-27T09:00:00Z"
                lines:
                  - lineNo: 1
                    sku: SKU-1
                    quantity: 2
                    pathId: pick
                    giftWrap: false
                    status: Released
                    reservationId: res-a1b2c3d4-0000-0000-0000-000000000009
                  - lineNo: 2
                    sku: SKU-2
                    quantity: 1
                    pathId: pick
                    giftWrap: true
                    status: Released
                    reservationId: res-a1b2c3d4-0000-0000-0000-000000000010
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/OrderNotFound"
        "409":
          description: This order has no backordered line to retry.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://errors.order-management.warehouse-systems.dev/no-backordered-lines
                title: Order has no backordered lines to retry
                status: 409
                detail: order has no backordered lines to retry
                instance: /orders/ord-a1b2c3d4-0000-0000-0000-000000000001/retry-allocation
        "503":
          $ref: "#/components/responses/DownstreamUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"
  /planned-capacity:
    get:
      operationId: getPlannedCapacity
      summary: List a site's planned capacity windows
      description: >
        Reads order-management's LOCAL read model of warehouse-planning's
        CapacityPlans for one site (ADR 0031): a PLANNED statement about a
        future window — the demand the planned capacity cannot process in
        it (`shortage`, in orders) — mirrored from the CloudEvents
        warehouse-planning publishes on `warehouse.warehouse-planning.events`.
        It is not live execution state and is never refreshed by calling
        warehouse-planning. Windows of every `status` are listed (a `DRAFT`
        is visible here but never influences an order; only a `PUBLISHED`
        plan with `shortage > 0` annotates orders — see `capacityConstraint`
        on the Order schema). Windows that ended before `from` are omitted.
        This route exists only when the integration is enabled; otherwise it
        is a 404.
      tags: [PlannedCapacity]
      parameters:
        - name: site
          in: query
          required: true
          description: >
            The planning `location` — a site/building code such as `SIM1`
            (exact, case-sensitive match).
          schema:
            type: string
            minLength: 1
          example: SIM1
        - name: from
          in: query
          required: false
          description: >
            RFC 3339 instant; only windows whose end is after it are listed.
            Defaults to now.
          schema:
            type: string
            format: date-time
          example: "2026-10-05T00:00:00Z"
      responses:
        "200":
          description: The site's planned windows ordered by start (possibly none).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannedCapacity"
              example:
                site: SIM1
                windows:
                  - planId: 0b7a4c1e-5d52-4f0e-9a39-6c1f2f3a8b10
                    warehouseId: WH-1
                    location: SIM1
                    pathId: pick-rebin-pack
                    windowStart: "2026-10-05T08:00:00Z"
                    windowEnd: "2026-10-05T16:00:00Z"
                    assignedDemand: 12000
                    capacityOverWindow: 8000
                    shortage: 4000
                    bottleneckStep: REBIN
                    status: PUBLISHED
                    asOf: "2026-10-04T21:45:10Z"
        "400":
          description: The `site` query parameter is missing, or `from` is not an RFC 3339 timestamp.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://errors.order-management.warehouse-systems.dev/invalid-query-parameter
                title: A required query parameter is missing or invalid
                status: 400
                detail: query parameter "site" is required
                instance: /planned-capacity
        "404":
          description: >
            The planned-capacity integration is not enabled on this
            deployment (`PLANNED_CAPACITY_CONSUMER_GROUP` unset): the route is
            not registered, so the router's default 404 answers.
        "500":
          $ref: "#/components/responses/InternalError"
components:
...
  responses:
    BadRequest:
      description: >
        The request body or path parameter is malformed or invalid —
        including a body whose JSON is syntactically valid but
        shape-invalid: an explicitly null value on a non-nullable field
        (e.g. `"giftWrap": null`) is rejected here rather than silently
        coerced to that field's zero value.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://errors.order-management.warehouse-systems.dev/empty-sku
            title: SKU must not be empty
            status: 400
            detail: sku must not be empty
            instance: /orders
...
  schemas:
...
    ReceiveOrderRequest:
      type: object
      required: [lines]
      properties:
        lines:
          type: array
          minItems: 1
          description: The requested lines. An order must have at least one.
          items:
            $ref: "#/components/schemas/ReceiveOrderLineRequest"
        requiredShipBy:
          type: string
          format: date-time
          description: >
            OPTIONAL. An externally-dictated deadline (RFC 3339) — the
            instant this order must ship by, set by someone other than this
            service (a retail network's required ship-by, for example).


            When present the promise is CONSTRAINED to a CPT window at or
            before it, rather than being the earliest window this service
            can manage; specifically, the LATEST window that still meets the
            deadline is chosen, leaving the floor maximum slack (ADR 0020
            §2). When the deadline cannot be met the order is still created
            and allocated, but comes back with NO `promiseDate` at all —
            an absent promise is the honest signal that we cannot commit,
            and is exactly what a caller holding a fill-or-kill obligation
            needs to see before acknowledging it downstream.


            Unlike the ordinary promise path, this one has NO lead-time
            fallback: missing schedule or capability data yields no promise
            rather than an estimate, because a guess dressed as a
            commitment is worse than an admission of ignorance.
        releaseOnAllocation:
          type: boolean
          default: true
          description: >
            When true (the default, and the behaviour of every caller that
            omits the field), the order is received, allocated and released
            in one call exactly as before ADR 0020. When false the order is
            HELD: lines are allocated, inventory reservations genuinely
            exist and a promise is computed and attached, but nothing is
            released and no work reaches the floor until
            `POST /orders/{id}/release`. Intended for a caller that must
            decide whether to commit to the order before it becomes work —
            see ADR 0020. A held order must be ship-complete: sending
            `releaseOnAllocation: false` together with
            `allowPartialShipment: true` is rejected with `422`.
        allowPartialShipment:
          type: boolean
          default: false
          description: >
            When false (the default — ship-complete, BR3), a single
            backordered line holds the WHOLE order back from release until
            retry-allocation clears it. When true, allocated lines are
            independently eligible for release.
          example: false
...